requirements.txt
flask==2.0.1
flask-socketio==5.1.1
python-dotenv==0.19.0
pandas==1.3.3
geopy==2.2.0
requests==2.26.0
python-dateutil==2.8.2
scikit-learn==1.3.0
schedule==1.2.0
tensorflow==2.15.0
keras==2.15.0
joblib==1.3.2 